Fixed some compile warnings. Updated todo.txt.
git-svn-id: file:///home/nick/svnrepo/repo/indigo/trunk@3945 20aa11e4-6e6f-594c-a032-6291ace9ee65

+++ physics/KDTree.cpp
@@ -1144,7 +1144,7 @@ void KDTree::saveTree(std::ostream& stream)
 	//write leafgeom
 	//------------------------------------------------------------------------
 	// Write number of leafgeom
-	temp = leafgeom.size();
+	temp = (unsigned int)leafgeom.size();
 	stream.write((const char*)&temp, sizeof(unsigned int));
 
 	// Write actual leafgeom data
